pyproject.toml
[tool.poetry]
name = "baldurcito"
version = "0.1.0"
description = ""
authors = ["jcrecio <juan.carlos.recio.abad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
huggingface-hub = "^0.20.3"
datasets = "^2.17.1"
pandas = "^2.2.0"
jsonlines = "^4.0.0"
bitsandbytes = "^0.43.0"
transformers = "^4.38.2"
peft = "^0.9.0"
accelerate = "^0.27.2"
trl = "^0.7.11"
torch = "2.2.0"
wandb = "^0.16.4"
einops = "^0.7.0"
sentencepiece = "^0.2.0"
poetry-dotenv-plugin = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
prepare_dataset = "baldurcito.prepare_dataset:main"